api/api/serializers/users.py
from rest_framework.serializers import ModelSerializer

from api.models import UserProfile


class UserProfileSerializer(ModelSerializer):
    class Meta:
        model = UserProfile
        fields = "__all__"
