api/pyproject.toml
[tool.poetry]
name = "fastcord"
version = "0.2.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvd.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
gunicorn = "^22.0.0"
uvicorn = "^0.30.3"
httpx = "^0.27.0"
markdown = "^3.6"
django-filter = "^24.2"
pyyaml = "^6.0.1"
uritemplate = "^4.1.1"
pygments = "^2.18.0"
django-guardian = "^2.4.0"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
dj-database-url = "^2.2.0"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
channels = {extras = ["daphne"], version = "^4.1.0"}
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"
pyjwt = "^2.8.0"
django-allauth = {extras = ["socialaccount"], version = "^0.63.6"}
cuid2 = "^2.0.1"
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
ignore-init-module-imports = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP"]
ignore = ["E203", "E266", "E501", "E402"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
